+++ R/globalVariables.R
@@ -1,11 +1,11 @@
 globalVariables(c('%>%', '.', 'DO_obs', 'DO_sat', 'Date', 'DateTimeStamp', 'Sal', 'Temp',
-                  'ats', 'bts', 'd', 'dDO', 'erts', 'exdat', 'gets', 'gppts', 'recompile', 
+                  'ats', 'bts', 'd', 'dDO', 'erts', 'nemts', 'exdat', 'gets', 'gppts', 'recompile', 
                   'update', 'Z', 'PAR', 'U10', 'a', 'b', 'jags.parallel', 'num.measurements',
-                  'pow', 'r', 'sc', 'nstepd', 'zz', 'D', 'P', 'R', 'matches', 'name',
+                  'pow', 'r', 'sc', 'nstepd', 'zz', 'D', 'P', 'R', 'NEM', 'matches', 'name',
                   'value', 'xval', 'grp', 'i', 'interval', 'isinterp', 'yval', 'DO_mod', 'rsq', 
-                  'atshi', 'atslo', 'btshi', 'btslo', 'ertshi', 'ertslo', 'getshi', 'getslo',
-                  'gpptshi', 'gpptslo', 'hi', 'lo', 'val', 'var', 'sd', 'uniobs', 'maxv',
-                  'cnt', 'ids', 'Rlo', 'Rhi', 'scaled'))
+                  'atshi', 'atslo', 'btshi', 'btslo', 'ertshi', 'ertslo', 'nemtslo', 'nemtshi', 
+                  'getshi', 'getslo', 'gpptshi', 'gpptslo', 'hi', 'lo', 'val', 'var', 'sd', 
+                  'uniobs', 'maxv', 'cnt', 'ids', 'Rlo', 'Rhi', 'scaled'))
 
 #' @importFrom stats lm na.omit runif
 NULL