+++ tests/testthat/test-ebase_plot.R
@@ -6,4 +6,12 @@ test_that("Checking ggplot class output, instantaneous", {
 test_that("Checking ggplot class output, daily", {
   result <- ebase_plot(exres, instantaneous = FALSE)
   expect_s3_class(result, 'ggplot')
+  ylb <- as.character(result$labels$y)
+  expect_true(grepl("Daily-averaged", ylb))
+})
+
+test_that("Checking ggplot class output, asnem true", {
+  result <- ebase_plot(exres, asnem = T)
+  expect_s3_class(result, 'ggplot')
+  expect_true("NEM" %in% result$data$name)
 })